refactor: SFM > rename tabpfn tradeable_regressor_manifest to tradeable_regressor
