+++ CHANGELOG.md
@@ -371,5 +371,5 @@ data mismatch
 - Refactor `lightgbm/tradeable_regressor` sfm to use manifest
 
 # V1.25.0 on 14th of December 2025
-- Add `sfm.tabpfn.tradeable_regressor_manifest` SFM using TabPFN model with manifest pattern
+- Add `sfm.tabpfn.tradeable_regressor` SFM using TabPFN model with manifest pattern
 - Add test for the above SFM
